// generated/api-client/models/facility-toggle-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FacilityToggleDto
 */
export interface FacilityToggleDto {
    /**
     * RFID card ID for facility access
     * @type {string}
     * @memberof FacilityToggleDto
     */
    'rfidId': string;
    /**
     * Facility ID to toggle access for
     * @type {string}
     * @memberof FacilityToggleDto
     */
    'facilityId': string;
}