// generated/api-client/models/facility-usage-status-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NestJS Authentication API
 * A comprehensive NestJS authentication API with PostgreSQL & Prisma.  ### Features - JWT Authentication with Access & Refresh tokens - Google OAuth2 integration - Local authentication (email/password) - Cookie-based token storage - Protected routes with guards  ### Authentication This API uses JWT tokens stored in HTTP-only cookies for authentication.  1. Local Authentication 2. Google OAuth
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FacilityUsageStatusDto
 */
export interface FacilityUsageStatusDto {
    /**
     * RFID card ID to check status
     * @type {string}
     * @memberof FacilityUsageStatusDto
     */
    'rfidId': string;
    /**
     * Facility ID to check status in
     * @type {string}
     * @memberof FacilityUsageStatusDto
     */
    'facilityId': string;
}